Move components into their own directory
