+++ src/application.js
@@ -1,6 +1,6 @@
 "use strict";
 
 var React = require('react');
-var Workspace = require('./workspace.jsx');
+var Workspace = require('./components/workspace.jsx');
 
 React.render(React.createElement(Workspace), document.getElementById('main'));
