+++ src/components/editor.jsx
